util-sheetprocessing/requirements.txt
gspread==6.1.4
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.4
